server/ai/requirements.txt
langchain==0.1.35
langchain-anthropic==0.0.13
langchain-openai==0.0.13
pypdf==4.0.1
faiss-cpu==1.7.4
openai==1.3.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
aioredis==2.0.1
redis==5.0.1
google-generativeai==0.3.0
pydantic==2.5.0
httpx==0.25.2
